build: simplify stack protector setup
The old configure code was written before the stack protector was
widely available and stable.  On modern systems, the compiler and
linker do the right thing if they support the flags at all.

Downgrade the protector from 'all' to 'strong' as the latter covers
any function taking the address of a stack object, providing good
protectiong with minimal overhead.  Also enable it unconditionally
when available.